requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
flask-jwt-extended==4.5.3
flask-cors==4.0.0
flask-bcrypt==1.0.1
email-validator==2.1.0
python-dotenv==1.0.0
marshmallow==3.20.1
pytest==7.4.2
pytest-flask==1.2.0
gunicorn==21.2.0
psycopg2-binary==2.9.9  # For PostgreSQL support

# Gmail API dependencies
google-api-python-client==2.88.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0

SQLAlchemy==2.0.20

mysqlclient==2.2.0
pymysql==1.1.0

sendgrid==6.10.0
gunicorn==21.2.0
bcrypt==4.0.1
Werkzeug==2.3.7
pytest==7.4.2